feat: create separate 7 modules
